+++ platformio/builder/tools/piolib.py
@@ -30,7 +30,7 @@ from SCons.Script import DefaultEnvironment  # pylint: disable=import-error
 
 from platformio import exception, fs
 from platformio.builder.tools import platformio as piotool
-from platformio.compat import IS_WINDOWS, hashlib_encode_data, string_types
+from platformio.compat import IS_WINDOWS, MISSING, hashlib_encode_data, string_types
 from platformio.http import HTTPClientError, InternetIsOffline
 from platformio.package.exception import (
     MissingPackageManifestError,
@@ -143,7 +143,7 @@ class LibBuilderBase:
 
         self._deps_are_processed = False
         self._circular_deps = []
-        self._processed_files = []
+        self._processed_search_files = []
 
         # reset source filter, could be overridden with extra script
         self.env["SRC_FILTER"] = ""
@@ -154,20 +154,27 @@ class LibBuilderBase:
     def __repr__(self):
         return "%s(%r)" % (self.__class__, self.path)
 
-    def __contains__(self, path):
-        p1 = self.path
-        p2 = path
+    def __contains__(self, child_path):
+        return self.is_common_builder(self.path, child_path)
+
+    def is_common_builder(self, root_path, child_path):
         if IS_WINDOWS:
-            p1 = p1.lower()
-            p2 = p2.lower()
-        if p1 == p2:
+            root_path = root_path.lower()
+            child_path = child_path.lower()
+        if root_path == child_path:
             return True
-        if os.path.commonprefix([p1 + os.path.sep, p2]) == p1 + os.path.sep:
+        if (
+            os.path.commonprefix([root_path + os.path.sep, child_path])
+            == root_path + os.path.sep
+        ):
             return True
         # try to resolve paths
-        p1 = os.path.os.path.realpath(p1)
-        p2 = os.path.os.path.realpath(p2)
-        return os.path.commonprefix([p1 + os.path.sep, p2]) == p1 + os.path.sep
+        root_path = os.path.realpath(root_path)
+        child_path = os.path.realpath(child_path)
+        return (
+            os.path.commonprefix([root_path + os.path.sep, child_path])
+            == root_path + os.path.sep
+        )
 
     @property
     def name(self):
@@ -324,7 +331,7 @@ class LibBuilderBase:
             )
         ]
 
-    def _get_found_includes(  # pylint: disable=too-many-branches
+    def get_implicit_includes(  # pylint: disable=too-many-branches
         self, search_files=None
     ):
         # all include directories
@@ -345,15 +352,17 @@ class LibBuilderBase:
         include_dirs.extend(LibBuilderBase._INCLUDE_DIRS_CACHE)
 
         result = []
-        for path in search_files or []:
-            if path in self._processed_files:
+        search_files = search_files or []
+        while search_files:
+            node = self.env.File(search_files.pop(0))
+            if node.get_abspath() in self._processed_search_files:
                 continue
-            self._processed_files.append(path)
+            self._processed_search_files.append(node.get_abspath())
 
             try:
                 assert "+" in self.lib_ldf_mode
                 candidates = LibBuilderBase.CCONDITIONAL_SCANNER(
-                    self.env.File(path),
+                    node,
                     self.env,
                     tuple(include_dirs),
                     depth=self.CCONDITIONAL_SCANNER_DEPTH,
@@ -363,39 +372,35 @@ class LibBuilderBase:
                 if self.verbose and "+" in self.lib_ldf_mode:
                     sys.stderr.write(
                         "Warning! Classic Pre Processor is used for `%s`, "
-                        "advanced has failed with `%s`\n" % (path, exc)
+                        "advanced has failed with `%s`\n" % (node.get_abspath(), exc)
                     )
-                candidates = self.env.File(path).get_implicit_deps(
-                    self.env,
-                    LibBuilderBase.CLASSIC_SCANNER,
-                    lambda _: tuple(include_dirs),
+                candidates = LibBuilderBase.CLASSIC_SCANNER(
+                    node, self.env, tuple(include_dirs)
                 )
 
-            # mark candidates already processed
-            self._processed_files.extend(
-                [
-                    c.get_abspath()
-                    for c in candidates
-                    if c.get_abspath() not in self._processed_files
-                ]
-            )
-
-            # print(path, [c.get_abspath() for c in candidates])
+            # print(node.get_abspath(), [c.get_abspath() for c in candidates])
             for item in candidates:
+                item_path = item.get_abspath()
+                # process internal files recursively
+                if (
+                    item_path not in self._processed_search_files
+                    and item_path not in search_files
+                    and item_path in self
+                ):
+                    search_files.append(item_path)
                 if item not in result:
                     result.append(item)
                 if not self.PARSE_SRC_BY_H_NAME:
                     continue
-                _h_path = item.get_abspath()
-                if not fs.path_endswith_ext(_h_path, piotool.SRC_HEADER_EXT):
+                if not fs.path_endswith_ext(item_path, piotool.SRC_HEADER_EXT):
                     continue
-                _f_part = _h_path[: _h_path.rindex(".")]
+                item_fname = item_path[: item_path.rindex(".")]
                 for ext in piotool.SRC_C_EXT + piotool.SRC_CXX_EXT:
-                    if not os.path.isfile("%s.%s" % (_f_part, ext)):
+                    if not os.path.isfile("%s.%s" % (item_fname, ext)):
                         continue
-                    _c_path = self.env.File("%s.%s" % (_f_part, ext))
-                    if _c_path not in result:
-                        result.append(_c_path)
+                    item_c_node = self.env.File("%s.%s" % (item_fname, ext))
+                    if item_c_node not in result:
+                        result.append(item_c_node)
 
         return result
 
@@ -410,7 +415,7 @@ class LibBuilderBase:
             search_files = self.get_search_files()
 
         lib_inc_map = {}
-        for inc in self._get_found_includes(search_files):
+        for inc in self.get_implicit_includes(search_files):
             inc_path = inc.get_abspath()
             for lb in self.env.GetLibBuilders():
                 if inc_path in lb:
@@ -571,11 +576,10 @@ class ArduinoLibBuilder(LibBuilderBase):
         # pylint: disable=no-member
         if not self._manifest.get("dependencies"):
             return LibBuilderBase.lib_ldf_mode.fget(self)
-        missing = object()
         global_value = self.env.GetProjectConfig().getraw(
-            "env:" + self.env["PIOENV"], "lib_ldf_mode", missing
+            "env:" + self.env["PIOENV"], "lib_ldf_mode", MISSING
         )
-        if global_value != missing:
+        if global_value != MISSING:
             return LibBuilderBase.lib_ldf_mode.fget(self)
         # automatically enable C++ Preprocessing in runtime
         # (Arduino IDE has this behavior)
@@ -827,11 +831,10 @@ class PlatformIOLibBuilder(LibBuilderBase):
 
     @property
     def lib_archive(self):
-        missing = object()
         global_value = self.env.GetProjectConfig().getraw(
-            "env:" + self.env["PIOENV"], "lib_archive", missing
+            "env:" + self.env["PIOENV"], "lib_archive", MISSING
         )
-        if global_value != missing:
+        if global_value != MISSING:
             return self.env.GetProjectConfig().get(
                 "env:" + self.env["PIOENV"], "lib_archive"
             )
@@ -881,6 +884,12 @@ class ProjectAsLibBuilder(LibBuilderBase):
         if export_projenv:
             env.Export(dict(projenv=self.env))
 
+    def __contains__(self, child_path):
+        for root_path in (self.include_dir, self.src_dir, self.test_dir):
+            if root_path and self.is_common_builder(root_path, child_path):
+                return True
+        return False
+
     @property
     def include_dir(self):
         include_dir = self.env.subst("$PROJECT_INCLUDE_DIR")
@@ -889,6 +898,10 @@ class ProjectAsLibBuilder(LibBuilderBase):
     @property
     def src_dir(self):
         return self.env.subst("$PROJECT_SRC_DIR")
+
+    @property
+    def test_dir(self):
+        return self.env.subst("$PROJECT_TEST_DIR")
 
     def get_search_files(self):
         items = []
@@ -1035,7 +1048,7 @@ def IsCompatibleLibBuilder(env, lb, verbose=int(ARGUMENTS.get("PIOVERBOSE", 0)))
             sys.stderr.write("Platform incompatible library %s\n" % lb.path)
         return False
     if compat_mode in ("soft", "strict") and not lb.is_frameworks_compatible(
-        env.get("PIOFRAMEWORK", [])
+        env.get("PIOFRAMEWORK", "__noframework__")
     ):
         if verbose:
             sys.stderr.write("Framework incompatible library %s\n" % lb.path)